backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-ollama
langchain-chroma
langchain-community
mem0ai
chromadb
neo4j
mcp
fastmcp
playwright
llama-index
markitdown
deepeval
ragas
instructor
guardrails-ai
langfuse
pydantic
pydantic-settings
python-dotenv
httpx
jinja2
python-multipart
tavily-python
